# insurance-app/services/application_service/app/models/applications.py
# -*- coding: utf-8 -*-
"""
保険契約申込に使用するレスポンスモデル定義
"""

# ------------------------------------------------------------------------------
# インポート
# ------------------------------------------------------------------------------
from pydantic import BaseModel, Field

# ------------------------------------------------------------------------------
# 出力モデル（バックエンド → フロントエンド）
# ------------------------------------------------------------------------------
class ApplicationResponseModel(BaseModel):
    """
    保険申込の応答モデル

    このモデルは、申込完了後にフロントエンドへ返却されるレスポンスとして使用されます。
    """
    application_id: str = Field(..., description="発行された申込ID（UUID形式）")
    quote_id: str = Field(..., description="申込元となった見積もりID")
    status: str = Field(..., description="申込ステータス（例：applied、pendingなど）")